Update Library.java
fixed the issues with the avail copies vs original stock number of books we have. also added comments for code that isn't working, or needs to be typed in correctly into the file so the brackets aren't off. 

code may be missing author names, please make sure to go in and add your name to what you worked on/ or helped debug/change.

what has been test: all the methods expect for load, save, and the main methods